indexing_coordinator/requirements.txt
requests
google-cloud-pubsub
protobuf

